Initial commit:
Created gradle Java Application with a 'backend' package
that will run the App. Other packages will then be structured
in a hexagonal way to make our app as dynamic as possible.

Uses Gradle as a package manager and Kotlin as the dsl language.

There is a first test as a proof on concept but will add
test metrics to encourage good test coverage.